Common tests
moved PaginateManagerTest into a tests directory in the common module.
